+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: tmarx <tmarx@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/04 14:11:17 by tom               #+#    #+#             */
-/*   Updated: 2019/10/04 14:13:00 by tom              ###   ########.fr       */
+/*   Updated: 2019/10/07 01:33:09 by tom              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char	*ft_strnstr(const char *str, const char *to_find, size_t len)
 	if (*to_find == '\0')
 		return ((char *)str);
 	i = 0;
-	while (str[i] && i < len)
+	while (str[i] && i < (int) len)
 	{
 		j = 0;
 		while (to_find[j] == str[i + j])
